safari-600.7-branch/LayoutTests/indieui/create-uirequestevent.html

<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Tests that document.createEvent() works with IndieUI events.');

var eventTypes = new Array("UIRequestEvent");

for (var k = 0; k < eventTypes.length; k++) {
    var event = document.createEvent("UIRequestEvent");

    debug(event);
    shouldBeTrue("typeof event == 'object'");

    shouldBeTrue("'type' in event");
    shouldBeTrue("'bubbles' in event");
    shouldBeTrue("'cancelable' in event");
    shouldBeTrue("'receiver' in event");

    shouldBe("typeof event.type", "'string'");
    shouldBe("typeof event.bubbles", "'boolean'");
    shouldBe("typeof event.cancelable", "'boolean'");
    shouldBe("typeof event.receiver", "'object'");

    debug("\n");
}



</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
