safari-600.7-branch/LayoutTests/indieui/create-uirequestevent-expected.txt

Tests that document.createEvent() works with IndieUI events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


[object UIRequestEvent]
PASS typeof event == 'object' is true
PASS 'type' in event is true
PASS 'bubbles' in event is true
PASS 'cancelable' in event is true
PASS 'receiver' in event is true
PASS typeof event.type is 'string'
PASS typeof event.bubbles is 'boolean'
PASS typeof event.cancelable is 'boolean'
PASS typeof event.receiver is 'object'


PASS successfullyParsed is true

TEST COMPLETE
